+++ scripts/prepare-icons.js
@@ -4,7 +4,7 @@ const png2icons = require('png2icons');
 
 (async () => {
   const source = path.join(__dirname, '../images/icon.png');
-  const destDir = path.join(__dirname, '../images/favicon.icns');
+  const destDir = path.join(__dirname, '../build/favicon.icns');
   await fs.ensureDir(destDir);
   const input = await fs.readFile(source);
   const output = png2icons.createICNS(input, png2icons.BICUBIC, 0);